requirements.txt
Django==4.2.6
djangorestframework==3.14.0
yfinance==0.2.31
django-admin-rangefilter=0.11.2